feat: render edit form via get request
Edit button on each table row has a hx-get to route + id. When one requests this, the server gets the contact via its id, and renders a slideout form component on the frontend.

We make sure that the original row is not swapped, and the button that calls edit has hx attributes to target the body and swap afterend